scaffolder-backend: separate out TemplateActionRegistry and types
